requirements.txt
asgiref==3.10.0
attrs==25.4.0
cattrs==25.2.0
dj-database-url==2.2.0
Django==5.2.7
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-graphql-jwt==0.4.0
django-schema-graph==3.1.0
django-timestampedmodel==0.1.0
graphene==3.4.3
graphene-django==3.2.3
graphene-django-cud==0.13.0
graphene-django-optimizer==0.10.0
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
packaging==25.0
promise==2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
schedule==1.2.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
time-stamped-model==0.2.11
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.6.0
